requirements.txt
# LangChain Agent 学习项目依赖

# 核心 LangChain 和 AI 库
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.12.0

# 数据处理和存储
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0

# Web 界面和 API
gradio>=4.15.0
fastapi>=0.108.0
uvicorn>=0.25.0

# 文档处理（为后续项目准备）
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0

# 数据科学（为数据分析项目准备）
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# 向量数据库（为 RAG 项目准备）
chromadb>=0.4.22
faiss-cpu>=1.7.4

# 开发和测试工具
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=23.12.0
flake8>=7.0.0

# 日志和监控
loguru>=0.7.2

# HTTP 请求
requests>=2.31.0
aiohttp>=3.9.0

# 实用工具
rich>=13.7.0
click>=8.1.0
tqdm>=4.66.0